fix(stella-model): rename the second reserved path, and build the gate that finds the rest in milliseconds
`crates/stella-model/src/credential/aux.rs` is the same defect the previous
commit fixed in `stella-cli`: `AUX` is a Windows reserved device name, git
refuses to write such a path, and `git checkout` exits 128 before anything
reaches disk. Two of them had been in this tree for as long as the modules
have existed, and nothing found them — the name is legal Rust and legal
POSIX, so no toolchain will ever report it.

`windows-check.yml` does find them, and finding the first one is how the
previous commit happened. But a failed checkout reports **one** bad path and
stops, so it is a detector with an O(n) feedback loop measured in CI minutes:
fix `stella-cli`'s, push, wait, learn about `stella-model`'s. That is the
wrong shape for a question that can be asked over the whole tracked set at
once, locally, with no toolchain.

`scripts/check-reserved-paths.sh` asks it: every path segment of every
tracked file, stem only, against AUX/CON/PRN/NUL/COM1-9/LPT1-9. It runs in
`GATE_GUARDS_FAST`, so `make guards-fast` and the pre-push hook cover it, and
it is the tracked set rather than the working tree because that is what a
clone reproduces. Checked against its own directions: a staged
`crates/stella-tools/src/aux.rs` makes it exit 1 naming that path, and
removing it makes it green again.

`make gate` is now 39 steps; AGENTS.md and CONTRIBUTING.md move with it, as
`check-gate-parity.sh` requires.

Refs #3550
