+++ crates/stella-model/src/credential/aux_credentials.rs
